перенесла функции в один файл
