+++ add_processor.php
@@ -1,6 +1,5 @@
 <?php
 
-
 $errors = validate_form($_POST);
 
 if (empty($errors)) {
